refactor: isolate three-proteome metrics module
